Added new APIs for indexes (#381)
Deprecated old index list and create methods.
Added new Builders for creating JSON and text index definitions.
Added common index type and specialised models for text and JSON
indexes.
Added associated Field types.
Added new Indexes class for listing multiple types.
Added new deleteIndex with type, redirected old deleteIndex to be for
JSON index only to preserve old behaviour.
Removed now unused internal code.
Added SelectorUtils for parsing selector strings.
Added custom serializers for field types and API asymmetry.
Added partial_filter_selector support.

Excluded findbugs false positive.

Updated IndexTests for new API.
Added mock tests for index create, list, delete.

Improved test helpers for new Index tests.
Created CloudantClientMockServerResource.
Updated DatabaseResource for mock server case.
Removed un-needed exception from signature in CloudantClientHelper.
Added mock tests for index creation, deletion and listing.
Added MockedServerTest class for easier mock tests.

Added index lifecycle test

Added CHANGES entry

Refactored UnicodeTest for new index API

+++ cloudant-client/src/test/java/com/cloudant/tests/util/DatabaseResource.java
@@ -34,6 +34,7 @@ public class DatabaseResource extends ExternalResource {
     private CloudantClient client;
     private String databaseName = Utils.generateUUID();
     private Database database;
+    private boolean mock = false;
 
     /**
      * Create a database resource from the specified clientResource. Note that if the resources
@@ -53,6 +54,11 @@ public class DatabaseResource extends ExternalResource {
      */
     public DatabaseResource(CloudantClientResource clientResource) {
         this.clientResource = clientResource;
+    }
+
+    public DatabaseResource(CloudantClientMockServerResource mockServerResource) {
+        this.clientResource = mockServerResource;
+        this.mock = true;
     }
 
     @Override
@@ -94,12 +100,18 @@ public class DatabaseResource extends ExternalResource {
     @Override
     public void before() {
         client = clientResource.get();
-        database = client.database(databaseName, true);
+        if (!mock) {
+            database = client.database(databaseName, true);
+        } else {
+            database = client.database(databaseName, false);
+        }
     }
 
     @Override
     public void after() {
-        client.deleteDB(databaseName);
+        if (!mock) {
+            client.deleteDB(databaseName);
+        }
     }
 
     public Database get() {
